refactored for contao 4
