+++ Resources/contao/config/autoload.ini
@@ -1,0 +1,14 @@
+;;
+; Configure what you want the autoload creator to register
+;;
+register_namespaces = false
+register_classes    = false
+register_templates  = false
+
+;;
+; Override the default configuration for certain sub directories
+;;
+[vendor/*]
+register_namespaces = false
+register_classes    = false
+register_templates  = false
